0, 1: fixed and understood header presedence |2: done

+++ 0x0E-function_pointers/1-array_iterator.c
@@ -1,10 +1,14 @@
-#include "function_pointers.h"
 #include <stdlib.h>
-#include <stddef.h>
-
+/**
+ * array_iterator - executes a function given as a parameter on
+ * each element of an array
+ * @array: list of ints
+ * @size: size of array of size_t (output of sizeof() and part of the stdlib.h)
+ * @action: pointer to function that prints int base 10 and base 16
+ */
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	size_t i;
+	unsigned int i;
 
 	if (array && action && size != 0)
 	{
